test: recovers unit tests up to a point (part 1)
Using injection of mocks and additional imports, the existing tests have been
recovered up to a point. They no longer fail because of missing providers
or imports. Many tests are still failing because conditions are not met. This, harder set of problems, will need to be closely evaluated test case by test case.